pyproject.toml
[tool.poetry]
name = "eocanvas"
version = "0.1.0"
description = "A Python wrapper to WEkEO Serverless Functions"
authors = ["Germano Guerrini <germano.guerrini@exprivia.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pyyaml = "^6.0.1"
lxml = "^5.2.2"
attrs = "^24.2.0"
Sphinx = { version = "6.2.1", optional = true }
sphinx-rtd-theme = { version = "2.0.0", optional = true }
sphinxcontrib-napoleon = { version = "0.7", optional = true }

[tool.poetry.extras]
docs = ["Sphinx", "sphinx-rtd-theme", "sphinxcontrib-napoleon"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
tox = "^4.15.0"
ipykernel = "^6.29.5"
responses = "^0.25.3"
pytype = "^2024.4.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 99
skip = [".tox", ".pytype"]

[tool.black]
line_length = 99
